+++ Source/CurveLooper.cpp
@@ -128,7 +128,7 @@ void CurveLooper::DrawModule()
 
    ofPushStyle();
    ofSetColor(ofColor::lime);
-   float x = ofLerp(mEnvelopeControl.GetPosition().x, mEnvelopeControl.GetPosition().x + mEnvelopeControl.GetDimensions().x, GetPlaybackPosition());
+   double x = ofLerp(mEnvelopeControl.GetPosition().x, mEnvelopeControl.GetPosition().x + mEnvelopeControl.GetDimensions().x, GetPlaybackPosition());
    ofLine(x, mEnvelopeControl.GetPosition().y, x, mEnvelopeControl.GetPosition().y + mEnvelopeControl.GetDimensions().y);
    ofPopStyle();
 }
