+++ Source/GroupControl.cpp
@@ -100,7 +100,7 @@ void GroupControl::CheckboxUpdated(Checkbox* checkbox, double time)
 
 namespace
 {
-   const float extraW = 20;
+   const double extraW = 20;
 }
 
 void GroupControl::GetModuleDimensions(double& width, double& height)
