+++ Source/LevelMeterDisplay.h
@@ -35,7 +35,7 @@ public:
    virtual ~LevelMeterDisplay();
 
    void Process(int channel, float* buffer, int bufferSize);
-   void Draw(float x, float y, float width, float height, int numChannels);
+   void Draw(double x, double y, double width, double height, int numChannels);
    void SetLimit(double limit);
 
 private:
