+++ Source/Panner.cpp
@@ -102,8 +102,8 @@ void Panner::Process(double time)
 
       ComputeSliders(i);
 
-      float left = GetBuffer()->GetChannel(0)[i];
-      float right = secondChannel[i];
+      double left = GetBuffer()->GetChannel(0)[i];
+      double right = secondChannel[i];
       GetBuffer()->GetChannel(0)[i] = left * ofMap(mPan, 0, 1, 1, 0, true) + right * ofMap(mPan, -1, 0, 1, 0, true);
       secondChannel[i] = right * ofMap(mPan, -1, 0, 0, 1, true) + left * ofMap(mPan, 0, 1, 0, 1, true);
